# [IMP] crm: improve _compute_meeting performance