--- a/addons/crm/models/res_partner.py
+++ b/addons/crm/models/res_partner.py
@@ -62,18 +62,19 @@ class Partner(models.Model):
     def _compute_meeting(self):
         if self.ids:
             all_partners = self.with_context(active_test=False).search([('id', 'child_of', self.ids)])
+
+            event_id = self.env['calendar.event']._search([])  # ir.rules will be applied
+            subquery_string, subquery_params = event_id.select()
+            subquery = self.env.cr.mogrify(subquery_string, subquery_params).decode()
+
             self.env.cr.execute("""
                 SELECT res_partner_id, calendar_event_id, count(1)
                   FROM calendar_event_res_partner_rel
-                 WHERE res_partner_id IN %s
+                 WHERE res_partner_id IN %s AND calendar_event_id IN ({})
               GROUP BY res_partner_id, calendar_event_id
-            """, [tuple(all_partners.ids)])
-            meeting_data = self.env.cr.fetchall()
+            """.format(subquery), [tuple(all_partners.ids)])
 
-            # Keep only valid meeting data based on record rules of events
-            events = [row[1] for row in meeting_data]
-            events = self.env['calendar.event'].search([('id', 'in', events)]).ids
-            meeting_data = [m for m in meeting_data if m[1] in events]
+            meeting_data = self.env.cr.fetchall()
 
             # Create a dict {partner_id: event_ids} and fill with events linked to the partner
             meetings = {p.id: set() for p in all_partners}