--- a/addons/crm/tests/__init__.py
+++ b/addons/crm/tests/__init__.py
@@ -9,3 +9,4 @@ from . import test_crm_lead_merge
 from . import test_crm_activity
 from . import test_crm_ui
 from . import test_crm_pls
+from . import test_res_partner