We can use ._search() to make sure that permissions and ir.rules are applied. By making use of psql subqueries, we remove the need of costly python filtering afterwards, thus increasing performance.

Benchmarks:

Running _compute_meeting on 1k partners in a local database.

| Number of active calendar_event | before PR | after PR |
|---|---|---|
| ~80k active calendar_event | 102.8200888633728 seconds | 0.7052993774414062 seconds |
| ~57k active calendar_event | 83.73815631866455 seconds | 0.5997936725616455 seconds |
| ~20k active calendar_event | 36.86744427680969 seconds | 0.3385133743286133 seconds |
| ~10k active calendar_event | 20.13909530639648 seconds | 0.2384305000305175 seconds |
| ~1k  active calendar_event | 4.009946823120117 seconds | 0.1153726577758789 seconds |


--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr