@tde-banana-odoo 2022-04-20
Just to know, did you try on Odoo db just to see the impact it could have ?