@awa-odoo 2022-04-20
Hello @dasa-odoo 

Would it be possible to include a small test that checks the meeting_count field value based on some crafted data?
Ideally it would include a small partner hierarchy and a few meetings on all of them.

Just to make sure we don't break anything while improving the performances.

(Make sure to double-check if such a test exists first but I did not find one).

Thanks for the PR & cheers!