@dasa-odoo 2022-04-25
@awa-odoo Hi there. I added the test you requested for meeting_count. Is it ok?

@tde-banana-odoo I didn't check on the Odoo db. Do I have access to that database?