@tde-banana-odoo 2022-04-26
@dasa-odoo Would be great to get in touch with internal team (tfr, nse, odo) and see with them if it is possible to test the patch on a test version of our db, just to have numbers on an heavily used db...