@tfr-odoo 2022-05-03
@tbe-odoo @dasa-odoo I have tested on staging4. It's now blazing fast

Before
```py
import time; st= time.time(); print(self.env['calendar.event'].search([], limit=10).mapped('partner_ids.meeting_count')); print(time.time() - st)
[22245, 7237, 18874, 4769, 9624, 34813, 4037]
63.02780604362488
```

After
```py
import time; st= time.time(); print(self.env['calendar.event'].search([], limit=10).mapped('partner_ids.meeting_count')); print(time.time() - st)
[22245, 7237, 18874, 4769, 9624, 34813, 4037]
4.386984348297119
```