@dasa-odoo 2022-05-06
@tde-banana-odoo I created a user and tested that ir.rules are applied.